Thu, Feb 20, 2025, 6:41 PM -03:00

+++ populaçao/populaçao.c
@@ -28,7 +28,7 @@ float anos_precisos = 0;
 
 if(populaçao_Atual < populaçao_Final)
 {
-   int  num_p = populaçao_Atual + (populaçao_Atual / 3) - (populaçao_Atual / 4);
+    float  num_p = populaçao_Atual + (populaçao_Atual / 3) - (populaçao_Atual / 4);
     anos_precisos++;
 }
 printf("Anos necessários para o aumento:%f  \n",anos_precisos);
